{"types":"C","runs":25}

+++ Problem30.java
@@ -6,7 +6,7 @@ public class Problem30 extends ConsoleProgram{
 		int n = readInt("enter a number: ");
 		while(true) {
 			int x = n / 10;
-			int k = n % 10;
+			double k = n % 10;
 			n = x;
 			print(k);
